[CBR-349] Pave the way for redemption
This cleans up some of the code a bit, stubs out some handlers for the
redemption end point.

As agreed with squad 1, this also moves the redemption endpoint to be part of
transactions endpoint rather than the accounts endpoint. This makes more sense,
since redemption creates a new transaction (otherwise it was the odd one out on
the accounts endpoint, the only one requiring additional parameters in the
legacy implementation, or requiring the active wallet in the new
implementation).